chore; lint fix and image rename
